Refactors Navigation component out of Page
