+++ src/components/page/page.module.css
@@ -4,40 +4,6 @@
   padding: 1rem .5rem;
 }
 
-@media (min-width: 30rem) {
-  .container {
-    padding: 1rem;
-  }
-}
-
-.identity {
-  display: flex;
-  gap: 1rem;
-  justify-content: center;
-  flex: 0 1 auto;
-}
-
-.avatar {
-  position: relative;
-  display: block;
-  overflow: hidden;
-  border-radius: 9999px;
-  flex: 0 0 auto;
-  align-self: center;
-  border: 1px solid var(--color-gray-100);
-  width: 3rem;
-  height: 3rem;
-  margin: 0;
-}
-
-.masthead {
-  flex: 0 1 8rem;
-}
-
-.masthead h1 {
-  margin: 0;
-}
-
 .title {
   font-family: var(--font-sans-serif);
   margin: 6rem 0 3rem;
@@ -45,94 +11,13 @@
   line-height: 1.3;
 }
 
-@media (min-width: 42rem) {
-  .sticky-top {
-    position: sticky;
-    top: 0;
-  }
-}
-
-.header {
+header.container {
   background: var(--page-color);
   width: 100vw;
   font-family: var(--font-sans-serif);
   padding: 2rem 0;
   margin-bottom: 0;
   z-index: 10;
-}
-
-.header .container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 2rem;
-  padding: .5rem;
-}
-
-.header img {
-  display: block;
-}
-
-.header h1 {
-  flex: 0 0 auto;
-  font-size: 1.25em;
-  font-weight: 700;
-  line-height: 1.2;
-  margin-bottom: .25rem;
-}
-
-.header p {
-  flex: 1 0 16rem;
-  margin: 0;
-  font-size: .85em;
-  line-height: 1.4;
-  font-weight: 400;
-}
-
-.header a {
-  color: var(--text-color);
-  text-decoration: none;
-  transition: opacity .1s linear;
-}
-
-.header a:hover {
-  opacity: .5;
-}
-
-.header a:visited {
-  color: var(--text-color);
-}
-
-.header .links {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem 2rem;
-}
-
-.header .links ul {
-  list-style-type: none;
-  display: flex;
-  gap: 1rem;
-  align-items: baseline;
-  font-weight: bold;
-  padding: 0;
-  margin: 0;
-}
-
-.icon {
-  display: block;
-  height: 1.2em;
-  width: 1.2em;
-  fill: var(--color-gray-700);
-}
-
-.githubIcon {
-  height: 1.4em;
-  width: 1.4em;
 }
 
 main.container {
@@ -151,15 +36,23 @@ main.container {
   z-index: 10;
 }
 
-@media (min-width: 36rem) {
-  .header .container {
-    flex-direction: row;
-    justify-content: space-between;
-    padding: .5rem 1rem;
+@media (min-width: 30rem) {
+  .container {
+    padding: 1rem;
   }
+}
 
-  .header .links {
-    flex-direction: row;
-    align-items: center;
+@media print {
+  .container {
+    margin: 1.5rem auto;
+  }
+  header.container {
+    padding: 0;
+  }
+  main.container {
+    padding: 0;
+  }
+  .envBanner {
+    display: none;
   }
 }
